modularization useWebsocketChat into smaller hooks
